pre create order

+++ src/main/java/com/deligo/Model/MenuItemInsert.java
@@ -2,10 +2,10 @@ package com.deligo.Model;
 
 public class MenuItemInsert {
     private int category_id;
-    private double price;
+    private float price;
     private boolean is_available;
 
-    public MenuItemInsert(int category_id, double price, boolean is_available) {
+    public MenuItemInsert(int category_id, float price, boolean is_available) {
         this.category_id = category_id;
         this.price = price;
         this.is_available = is_available;
@@ -23,7 +23,7 @@ public class MenuItemInsert {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
